Split virtio.c in virtio_net and virtio_blk
